Print env change shell code on stdout
If a shell name is given as argument (among sh, bash, ksh, zsh, csh,
tcsh and fish), print shell code for this shell kind on stdout.

Printed code corresponds to the environment changes induced by the
module actions performed on the GUI.

With this new option, if a mogui() shell function is defined and
evaluates output produced by mogui.py script, the environment changes
are reflected in calling shell session. This is the same mechanism as
used by module() shell function to apply environment changes produced by
modulecmd.tcl.