Report display info when right clicking  loaded modules view
Add show_display member to LoadedModulesView class. It defines the
function to call to show display info. This member is set to a method of
MoGui class that calls module help command for selected module item (new
method display() added to Module class).

LoadedModulesView widget is updated to create its items as ModuleGui
items (as done on AvailModulesView widget). It helps to proceed with the
report of display information like done with the report of help
information on avail modules.